Создаёт контроллер экрана «Фильмы»
